# perf(script): memoize dynamic function