--- a/lib/script.js
+++ b/lib/script.js
@@ -20,7 +20,7 @@ Script.prototype.runWithContext = function (context) {
   functionArgs.splice(functionArgs.indexOf('this'), 1);
 
   functionArgs.push(this.scriptSourceCode);
-  var func = Function.apply(null, functionArgs);
+  var func = this.getFunction(functionArgs);
 
   // pass our arguments from the sandbox to the function
   var args = [];
@@ -29,6 +29,10 @@ Script.prototype.runWithContext = function (context) {
   });
   return func.apply(context._this || {}, args);
 };
+
+Script.prototype.getFunction = _.memoize(function(functionArgs) {
+  return Function.apply(null, functionArgs);
+});
 
 /**
  * Run the current script in the given sandbox. An optional domain may be provided to extend the sandbox exposed to the script.